requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
PyYAML==6.0.2
Pillow==12.1.1
plotly
kaleido
pytest==8.0.0
